requirements.txt
requests
python-dotenv
logging
pyarrow
psycopg2-binary
pandas
numpy
boto3